// frontend/todo-app/src/api/todo-api.ts
export interface TodoResponseDto {
  /** @format int32 */
  id?: number;
  name?: string | null;
  description?: string | null;
  /** @format date-time */
  forDate?: string;
  isCompleted?: boolean;
  /** @format date-time */
  createdAt?: string;
  /** @format date-time */
  modifiedAt?: string;
}